pyproject.toml
[project]
name = "bright-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.4.0",
    "brightdata-sdk>=2.1.0",
    "duckdb>=1.4.3",
    "fastexcel>=0.18.0",
    "google-genai>=1.57.0",
    "google-generativeai>=0.8.6",
    "pandas>=2.3.3",
    "plotly>=6.5.1",
    "polars>=1.37.0",
    "pyarrow>=22.0.0",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "xlsxwriter>=3.2.9",
    "bcrypt>=4.2.0",
    "passlib>=1.7.4",
    # Worker Dependencies
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "openpyxl>=3.1.5",
    "playwright>=1.57.0",
]
